Move the file reader writer to a files package.
