+++ pkg/commands/cmdlib/BUILD.bazel
@@ -13,7 +13,7 @@ go_library(
     deps = [
         "//pkg/apis/bundle/v1alpha1:go_default_library",
         "//pkg/converter:go_default_library",
-        "//pkg/core:go_default_library",
+        "//pkg/files:go_default_library",
         "//pkg/transformer:go_default_library",
         "//vendor/github.com/ghodss/yaml:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
