+++ pkg/commands/export/BUILD.bazel
@@ -12,7 +12,7 @@ go_library(
         "//pkg/apis/bundle/v1alpha1:go_default_library",
         "//pkg/commands/cmdlib:go_default_library",
         "//pkg/converter:go_default_library",
-        "//pkg/core:go_default_library",
+        "//pkg/files:go_default_library",
         "//pkg/transformer:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
