+++ pkg/commands/find/images.go
@@ -19,7 +19,7 @@ import (
 	"fmt"
 
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/commands/cmdlib"
-	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/core"
+	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/files"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/find"
 	log "github.com/golang/glog"
 	"github.com/spf13/cobra"
@@ -31,14 +31,14 @@ type options struct {
 var opts = &options{}
 
 func findAction(ctx context.Context, cmd *cobra.Command, _ []string) {
-	rw := &core.LocalFileSystemReaderWriter{}
+	rw := &files.LocalFileSystemReaderWriter{}
 	gopts := cmdlib.GlobalOptionsValues.Copy()
 	if err := runFindImages(ctx, opts, rw, gopts); err != nil {
 		log.Exitf("error in runFindImages: %v", err)
 	}
 }
 
-func runFindImages(ctx context.Context, _ *options, rw core.FileReaderWriter, gopt *cmdlib.GlobalOptions) error {
+func runFindImages(ctx context.Context, _ *options, rw files.FileReaderWriter, gopt *cmdlib.GlobalOptions) error {
 	b, err := cmdlib.ReadBundleContents(ctx, rw, gopt)
 	if err != nil {
 		return fmt.Errorf("error reading bundle contents: %v", err)
