+++ pkg/commands/modify/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/commands/cmdlib:go_default_library",
-        "//pkg/core:go_default_library",
+        "//pkg/files:go_default_library",
         "//pkg/transformer:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
