+++ pkg/core/BUILD.bazel
@@ -5,12 +5,8 @@ go_library(
     srcs = [
         "core.go",
         "doc.go",
-        "file_reader_writer.go",
     ],
     importpath = "github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/core",
     visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/apis/bundle/v1alpha1:go_default_library",
-        "//vendor/github.com/golang/protobuf/ptypes/struct:go_default_library",
-    ],
+    deps = ["//vendor/github.com/golang/protobuf/ptypes/struct:go_default_library"],
 )
