+++ pkg/files/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = [
+        "doc.go",
+        "file_reader_writer.go",
+    ],
+    importpath = "github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/files",
+    visibility = ["//visibility:public"],
+    deps = ["//pkg/apis/bundle/v1alpha1:go_default_library"],
+)
