+++ pkg/files/doc.go
@@ -1,0 +1,2 @@
+// Package files contains methods and data types for manipulating files.
+package files
