+++ pkg/transformer/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
         "//pkg/apis/bundle/v1alpha1:go_default_library",
         "//pkg/converter:go_default_library",
         "//pkg/core:go_default_library",
+        "//pkg/files:go_default_library",
         "//pkg/find:go_default_library",
         "//vendor/github.com/golang/protobuf/proto:go_default_library",
     ],
